requirements.txt
chainlit>=1.3.0
langgraph>=0.2.50
langgraph-checkpoint-sqlite>=1.0.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langchain-google-community>=0.1.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
duckduckgo-search>=6.3.0
requests>=2.31.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
python-docx>=1.1.0
pypdf>=4.0.0
mammoth>=1.6.0
anthropic>=0.40.0
feedparser>=6.0.10
